requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2

# GCP
google-cloud-tasks==2.19.3

# Supabase + HTTP
supabase==2.6.0
httpx==0.27.2

# PDF
jinja2==3.1.4
weasyprint==61.2

# Gemini (official client)
google-generativeai==0.7.2

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.2  # Already included above, needed for testing